Filtering, sorting and empty message updated
